Removing brute force redo of lobby data every time someone joins/changes info
Unfortunately, in my attempt to make this lobby player data more type safe, I ran into an interesting issue with Godot: you can't pass objects unless they're one of the "primitive" supported types like Dictionaries.

It turns out you also get a cyclical dependency error if you make a factory method on a class that generates instances of the class, so things got ugly.

For now, I made some DTO conversion logic and split out the class because at least then we can restrict the type unsafe code to the inner workings of the Network class... but we'll see how we feel about it in the morning.  I feel pretty dirty about it.