Created new Timer class
Timers are not framerate independent, this caused lots of problems when trying to synchronize things across the network. Slower clients would timeout way later than other clients. WallClockTimer will now timeout at the same time no matter what. It is also a drop in replacement for the Timer class. In the future, the Server can send a Unix timestamp for all clients to use for their timer. This final step should ultimately fix clients not starting the game at the same time.